requirements.txt
characterai
vk-api
whatismyip
typing_extensions
websockets
pyyaml